# SavePoint/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SavePoint API",
    "version": "1.0.0",
    "description": "API REST do SavePoint: catálogo de jogos, comparação de preços entre lojas, histórico de preços, avaliações de usuários e lista de desejos."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/lojas": {
      "get": {
        "tags": [
          "Lojas"
        ],
        "summary": "Listar todas as lojas",
        "operationId": "loja_listarTodas",
        "responses": {
          "200": {
            "description": "Lista de lojas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loja"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lojas"
        ],
        "summary": "Cadastrar loja",
        "operationId": "loja_salvar",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Loja"
              },
              "example": {
                "nome": "Steam",
                "urlLoja": "https://store.steampowered.com",
                "urlLogo": "https://exemplo.com/logos/steam.png"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loja"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou nome já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/lojas/{id}": {
      "get": {
        "tags": [
          "Lojas"
        ],
        "summary": "Buscar loja por ID",
        "operationId": "loja_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loja"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Lojas"
        ],
        "summary": "Atualizar loja",
        "operationId": "loja_atualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Loja"
              },
              "example": {
                "nome": "Steam",
                "urlLoja": "https://store.steampowered.com",
                "urlLogo": "https://exemplo.com/logos/steam.png"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Loja atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loja"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou nome já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Lojas"
        ],
        "summary": "Excluir loja",
        "operationId": "loja_deletar",
        "description": "Remove também as ofertas da loja e o histórico delas.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removido com sucesso (sem conteúdo)"
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/jogos/{id}/ofertas": {
      "post": {
        "tags": [
          "Ofertas"
        ],
        "summary": "Cadastrar oferta de um jogo em uma loja",
        "operationId": "ofertaJogo_cadastrarOferta",
        "description": "O percentual de desconto é calculado a partir dos preços original e atual.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfertaJogo"
              },
              "example": {
                "precoOriginal": 199.9,
                "precoAtual": 99.9,
                "loja": {
                  "id": 1
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaJogo"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Jogo ou loja não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/ofertas/{id}/comparacao": {
      "get": {
        "tags": [
          "Ofertas"
        ],
        "summary": "Comparar preços do jogo entre lojas",
        "operationId": "ofertaJogo_compararPrecos",
        "description": "O parâmetro de caminho é o ID do jogo.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ofertas ordenadas do menor para o maior preço atual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OfertaJogo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Jogo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/ofertas/{id}": {
      "get": {
        "tags": [
          "Ofertas"
        ],
        "summary": "Buscar oferta por ID",
        "operationId": "ofertaJogo_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaJogo"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Loja": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "urlLoja": {
            "type": "string"
          },
          "urlLogo": {
            "type": "string"
          }
        }
      },
      "OfertaJogo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "precoOriginal": {
            "type": "number",
            "format": "double"
          },
          "precoAtual": {
            "type": "number",
            "format": "double"
          },
          "percentualDesconto": {
            "type": "number",
            "format": "float"
          },
          "loja": {
            "$ref": "#/components/schemas/Loja"
          },
          "historico": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Historico"
            }
          }
        }
      },
      "Historico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "preco": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "ErroResposta": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "erro": {
            "type": "string"
          },
          "mensagem": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}